+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add success and error types for activity logs
 - Include status indicator for each activity log on the dashboard
+- Add an activity status filter to the dashboard
 
 ## 1.0.0-beta.1 - 2024-02-20